WIP - implementing parser
This will become a functioning server which can accept requests, but the server aspect is not doing much yet.

The focus atm is on the parser. This contains the basic class structure for a grammar, but is pretty feature light (as in, it doesn't do anything). Next steps:
1. Compilation of grammars to CNF
2. Parsing of strings with compiled grammars
3. Evaluating parsed strings

All of those steps are pretty nontrivial and this is a spare time project, so we'll see!